refactor: move some functions to widgets
